Changes after review, added Arbeidsforhold summary component
