+++ src/views/Summary.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { QuizFaktum, QuizSeksjon } from "../types/quiz.types";
+import { QuizSeksjon } from "../types/quiz.types";
 import { Accordion } from "@navikt/ds-react";
-import { SummaryFaktum } from "../components/summary/SummaryFaktum";
 import { ARBEIDSFORHOLD_FAKTUM_ID } from "../constants";
-import { getArbeidsforholdName } from "../components/arbeidsforhold/Arbeidsforhold";
+import { Arbeidsforhold } from "../components/summary/Arbeidsforhold";
+import { Faktum } from "../components/faktum/Faktum";
 
 interface Props {
   section: QuizSeksjon[];
@@ -23,47 +23,19 @@ export function Summary(props: Props) {
                     {section.fakta
                       ?.filter((faktum) => {
                         // Arbeidsforhold vises som en egen accordion
-                        return faktum.type !== "generator";
+                        return faktum.beskrivendeId !== ARBEIDSFORHOLD_FAKTUM_ID;
                       })
                       .map((faktum) => {
-                        return <SummaryFaktum key={faktum.id} faktum={faktum} />;
+                        return <Faktum key={faktum.id} faktum={faktum} readonly={true} />;
                       })}
                   </>
                 </Accordion.Content>
               </Accordion.Item>
-              {visArbeidsforhold(section)}
+              <Arbeidsforhold section={section} />
             </>
           );
         })}
       </Accordion>
     </>
   );
-}
-
-function visArbeidsforhold(section: QuizSeksjon) {
-  {
-    const arbeidsforholdSvar = section.fakta?.find(
-      (faktum) => faktum.beskrivendeId === ARBEIDSFORHOLD_FAKTUM_ID
-    )?.svar as QuizFaktum[][];
-
-    if (arbeidsforholdSvar && arbeidsforholdSvar.length > 0) {
-      return arbeidsforholdSvar.map(function (arbeidsforhold, index) {
-        const tittel = getArbeidsforholdName(arbeidsforhold);
-        const id = `${section.beskrivendeId}-${index}`;
-
-        return (
-          <Accordion.Item key={id}>
-            <Accordion.Header>{tittel}</Accordion.Header>
-            <Accordion.Content>
-              <>
-                {arbeidsforhold.map((faktum) => {
-                  return <SummaryFaktum key={faktum.id} faktum={faktum} />;
-                })}
-              </>
-            </Accordion.Content>
-          </Accordion.Item>
-        );
-      });
-    }
-  }
 }
